fix 3.3.25
