+++ example/example-server/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>springboot-example</artifactId>
         <groupId>com.codingapi.springboot</groupId>
-        <version>3.3.24</version>
+        <version>3.3.25</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
