+++ springboot-starter-security/pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <artifactId>springboot-parent</artifactId>
         <groupId>com.codingapi.springboot</groupId>
-        <version>3.3.24</version>
+        <version>3.3.25</version>
     </parent>
 
     <artifactId>springboot-starter-security</artifactId>
