+++ springboot-starter/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>com.codingapi.springboot</groupId>
         <artifactId>springboot-parent</artifactId>
-        <version>3.3.24</version>
+        <version>3.3.25</version>
     </parent>
     <artifactId>springboot-starter</artifactId>
